+++ test/parallel/test-readline-interface.js
@@ -869,3 +869,80 @@ function isWarned(emitter) {
     assert.strictEqual(rl._prompt, '$ ');
   }
 });
+
+// For the purposes of the following tests, we do not care about the exact
+// value of crlfDelay, only that the behaviour conforms to what's expected.
+// Setting it to Infinity allows the test to succeed even under extreme
+// CPU stress.
+const crlfDelay = Infinity;
+
+[ true, false ].forEach(function(terminal) {
+  // sending multiple newlines at once that does not end with a new line
+  // and a `end` event(last line is)
+
+  // \r\n should emit one line event, not two
+  {
+    const fi = new FakeInput();
+    const rli = new readline.Interface(
+      {
+        input: fi,
+        output: fi,
+        terminal: terminal,
+        crlfDelay
+      }
+    );
+    const expectedLines = ['foo', 'bar', 'baz', 'bat'];
+    let callCount = 0;
+    rli.on('line', function(line) {
+      assert.strictEqual(line, expectedLines[callCount]);
+      callCount++;
+    });
+    fi.emit('data', expectedLines.join('\r\n'));
+    assert.strictEqual(callCount, expectedLines.length - 1);
+    rli.close();
+  }
+
+  // \r\n should emit one line event when split across multiple writes.
+  {
+    const fi = new FakeInput();
+    const rli = new readline.Interface({
+      input: fi,
+      output: fi,
+      terminal: terminal,
+      crlfDelay
+    });
+    const expectedLines = ['foo', 'bar', 'baz', 'bat'];
+    let callCount = 0;
+    rli.on('line', function(line) {
+      assert.strictEqual(line, expectedLines[callCount]);
+      callCount++;
+    });
+    expectedLines.forEach(function(line) {
+      fi.emit('data', `${line}\r`);
+      fi.emit('data', '\n');
+    });
+    assert.strictEqual(callCount, expectedLines.length);
+    rli.close();
+  }
+
+  // Emit one line event when the delay between \r and \n is
+  // over the default crlfDelay but within the setting value.
+  {
+    const fi = new FakeInput();
+    const delay = 125;
+    const rli = new readline.Interface({
+      input: fi,
+      output: fi,
+      terminal: terminal,
+      crlfDelay
+    });
+    let callCount = 0;
+    rli.on('line', () => callCount++);
+    fi.emit('data', '\r');
+    setTimeout(common.mustCall(() => {
+      fi.emit('data', '\n');
+      assert.strictEqual(callCount, 1);
+      rli.close();
+    }), delay);
+  }
+});
